refactor(wl): create server connections in free function
This allows external reuse.
